Version 0.2

+++ src/main/java/randomazer/repos/NumericRepo.java
@@ -3,5 +3,5 @@ package randomazer.repos;
 import org.springframework.data.repository.CrudRepository;
 import randomazer.domain.Numeric;
 
-public interface NumericRepo extends CrudRepository<Numeric, Integer> {
+public interface NumericRepo extends CrudRepository<Numeric, Long> {
 }
